calculating floor earned money
Pretty much works. Only issue lies in cases where halls Blum, Mees, or
Scharp having an attendence of >= 2 but <4 for a meeting and being
incorrectly counted as having attended. (These halls require 4 students
to attend before being counted as having attended as a hall). Will be
using the FloorAttendanceNumerics table for determining this step.