fixed some issues
Fixed spelling error in several functions in adminPermissions.js, and
fixed the issue in signUps.js that prevented all of the functionality
from properly displaying for officers even if they were logged in. Issue
ended up having to do with load times, where the webpage would load
faster than the officer XHR request could send a response and inform the
page that the user currently logged in had access to that information.