Adding/Deleting Equipment
Created a new button that allows an admin to add a new piece of
equipment. Assumes that the calendar already exists and requests the
embeded text in the modal. Also added a "delete" button that draws
whenever a calendar is selected. Clicking this and the subsequent popup
delete button removes the calendar/equipment from the database.