moved some code around
now calls members/ endpoint instead, and added an addtional check to
ensure that the officer-specific forloop is still checking what we
consider officers only. This lets me also check every member
simultaneously so I dont have to make additional requests to determine
if someone has been added to the database yet or not.